Get rid of a compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@1252 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogPromptDlg.cpp
@@ -1281,7 +1281,7 @@ void CLogPromptDlg::OnEnChangeLogmessage()
 {
 	CString sTemp;
 	GetDlgItem(IDC_LOGMESSAGE)->GetWindowText(sTemp);
-	if (sTemp.GetLength() > CRegDWORD(_T("Software\\TortoiseSVN\\MinLogSize"), 10))
+	if (DWORD(sTemp.GetLength()) > CRegDWORD(_T("Software\\TortoiseSVN\\MinLogSize"), 10))
 	{
 		GetDlgItem(IDOK)->EnableWindow(TRUE);
 	}
